Moving top-tagging stuff from Analysis domain to RECO domain, and implementing some substructure plots into the B2GDQM module
