+++ AnalysisDataFormats/TopObjects/src/classes_def.xml
@@ -65,21 +65,6 @@
   <class name="edm::Wrapper<std::vector<TtHadEvtSolution> >" />
   <class name="edm::Wrapper<std::vector<StEvtSolution> >" />
 
-  <class name="reco::CATopJetProperties" ClassVersion="10">
-   <version ClassVersion="10" checksum="4096516518"/>
-  </class>
-  <class name="std::pair<edm::RefToBase<reco::Jet>, reco::CATopJetProperties>"/>
- 
-  <class name="reco::CATopJetTagInfo" ClassVersion="10">
-   <version ClassVersion="10" checksum="3059833435"/>
-  </class>
-  <class name="reco::CATopJetTagInfoCollection"/>
-  <class name="reco::CATopJetTagInfoRef"/>
-  <class name="reco::CATopJetTagInfoRefProd"/>
-  <class name="reco::CATopJetTagInfoRefVector"/>
-  <class name="edm::Wrapper<reco::CATopJetTagInfoCollection>"/>
-  <class name="edm::reftobase::Holder<reco::BaseTagInfo, reco::CATopJetTagInfoRef>" />
-  <class name="edm::reftobase::RefHolder<reco::CATopJetTagInfoRef>" />
 
 
 </lcgdict>
