+++ RecoJets/JetProducers/plugins/CATopJetTagger.h
@@ -50,7 +50,6 @@
 #include "DataFormats/JetReco/interface/BasicJet.h"
 #include "DataFormats/JetReco/interface/CaloJet.h"
 #include "DataFormats/Candidate/interface/CompositeCandidate.h"
-#include "PhysicsTools/CandUtils/interface/AddFourMomenta.h"
 #include "DataFormats/Candidate/interface/CandMatchMap.h"
 
 #include <Math/VectorUtil.h>
